Wall width change, rotating aimer

+++ Server/ConfigManager.h
@@ -77,9 +77,9 @@ public:
   static float projectileRange(){ return StringToNumber<float>(ConfigManager::configMap["projectileRange"]); }
 
   /// Wall ///
-  static unsigned int wallDepth(){ return StringToNumber<unsigned int>(ConfigManager::configMap["wallDepth"]); }
-  static unsigned int wallHeight(){ return StringToNumber<unsigned int>(ConfigManager::configMap["wallHeight"]); }
-  static unsigned int wallWidth(){ return StringToNumber<unsigned int>(ConfigManager::configMap["wallWidth"]); }
+  static float wallDepth(){ return StringToNumber<float>(ConfigManager::configMap["wallDepth"]); }
+  static float wallHeight(){ return StringToNumber<float>(ConfigManager::configMap["wallHeight"]); }
+  static float wallWidth(){ return StringToNumber<float>(ConfigManager::configMap["wallWidth"]); }
 
   /// Physics ///
   static float gravity(){  return StringToNumber<float>(ConfigManager::configMap["gravity"]); }
